Add files via upload
The LCD portion works but I can't get the Keypad to work correctly.  It should be pretty straight forward but it doesn't seem to work.  Can one of you take a look.  If you look at the main function while(1) loop, the commented out part is the issue.  If you uncomment it and comment out the display test, it should work but doesn't.